Extract common base class for RulerToggle*Breakpoint* delegates
See https://github.com/eclipse-jdt/eclipse.jdt.debug/pull/660
